Enabled observability and modified observability mechanism
