+++ .gitignore
@@ -12,3 +12,4 @@ src/report*.json
 kubearmor_policies*.yaml
 cilium_policies*.yaml
 knox_net_policies*.yaml
+src/accuknox.db
