+++ src/conf/local-file.yaml
@@ -27,7 +27,7 @@ application:
     #cluster-mgmt-url: "http://cluster-management-service.accuknox-dev-cluster-mgmt.svc.cluster.local/cm"
     cluster-mgmt-url: "http://localhost:8080"
 
-observability: false
+observability: true
 
 database:
   driver: sqlite3
