+++ src/systempolicy/systemPolicy.go
@@ -19,6 +19,7 @@ import (
 	"github.com/accuknox/auto-policy-discovery/src/feedconsumer"
 	"github.com/accuknox/auto-policy-discovery/src/libs"
 	logger "github.com/accuknox/auto-policy-discovery/src/logging"
+	"github.com/accuknox/auto-policy-discovery/src/observability"
 	"github.com/accuknox/auto-policy-discovery/src/plugin"
 	wpb "github.com/accuknox/auto-policy-discovery/src/protobuf/v1/worker"
 	types "github.com/accuknox/auto-policy-discovery/src/types"
@@ -1366,6 +1367,13 @@ func StartSystemCronJob() {
 	if err != nil {
 		log.Error().Msg(err.Error())
 		return
+	}
+	if cfg.CurrentCfg.Observability {
+		err := SystemCronJob.AddFunc(cfg.GetCfgSysCronJobTime(), observability.SystemLogCronJob) // time interval
+		if err != nil {
+			log.Error().Msg(err.Error())
+			return
+		}
 	}
 	SystemCronJob.Start()
